+++ cockpit/core/config/system_config.h
@@ -89,6 +89,15 @@ struct CloudUplinkConfig {
   MqttConfig mqtt;
 };
 
+struct RecordingServiceConfig {
+  bool auto_start = false;
+  std::string directory = "recordings";
+  std::string vehicle_data_address = "127.0.0.1:50050";
+  int stream_timeout_ms = 10000;
+  int retry_delay_ms = 200;
+  GrpcServerConfig grpc{"127.0.0.1:50055"};
+};
+
 struct ServicesConfig {
   VehicleDataConfig vehicle_data;
   GatewayConfig gateway;
@@ -96,6 +105,7 @@ struct ServicesConfig {
   CameraServiceConfig camera;
   VoiceInteractionServiceConfig voice_interaction;
   CloudUplinkConfig cloud_uplink;
+  RecordingServiceConfig recording;
 };
 
 struct CanConfig {